lot of housecleaning on jest tests
