+++ test/js/bun/test/test-test.test.ts
@@ -1586,1319 +1586,6 @@ test("testing Bun.deepEquals() using isEqual()", () => {
   expect(-Infinity).toEqual(-1 / 0);
 });
 
-test("toHaveProperty() - emojis", () => {
-  expect({ "👍": "thumbs up" }).toHaveProperty("👍", "thumbs up");
-  expect({ "👩‍👩‍👧‍👧": "family" }).toHaveProperty("👩‍👩‍👧‍👧", "family");
-  expect({ "😶‍🌫️": "fog" }).toHaveProperty("😶‍🌫️", "fog");
-  expect({ "👩‍❤️‍👨": "couple" }).toHaveProperty("👩‍❤️‍👨", "couple");
-  expect({ "👩‍❤️‍👨‍👨‍👧‍👧": "family" }).toHaveProperty("👩‍❤️‍👨‍👨‍👧‍👧", "family");
-  expect({ "👩‍❤️‍👨‍👨‍👧": "family" }).toHaveProperty("👩‍❤️‍👨‍👨‍👧", "family");
-  expect({ "👩‍❤️‍👨‍👨‍👧": "family" }).not.toHaveProperty("👩‍❤️‍👨‍👨‍👧‍👧", "family");
-
-  // emojis in array
-  expect(["👍", "👎"]).toHaveProperty("0", "👍");
-  expect(["👍", "👎"]).toHaveProperty("1", "👎");
-  expect(["👍", "👎"]).not.toHaveProperty("0", "👎");
-  expect(["👍", "👎"]).not.toHaveProperty("1", "👍");
-  expect(["👩‍❤️‍👨‍👨‍👧‍👧"]).toHaveProperty("0", "👩‍❤️‍👨‍👨‍👧‍👧");
-  expect(["👩‍❤️‍👨‍👨‍👧‍👧"]).toHaveProperty([0], "👩‍❤️‍👨‍👨‍👧‍👧");
-  expect(["😶‍🌫️"]).toHaveProperty([0], "😶‍🌫️");
-});
-
-test("toHaveProperty() - dot and bracket notation edge cases", () => {
-  expect({ a: 1 }).not.toHaveProperty(".");
-  expect({ a: 1 }).not.toHaveProperty("]");
-  expect({ a: 1 }).not.toHaveProperty("[");
-  expect({ a: 1 }).not.toHaveProperty("[]");
-  expect({ a: 1 }).not.toHaveProperty("[[]]");
-  expect({ a: 1 }).not.toHaveProperty("[[");
-  expect({ a: 1 }).not.toHaveProperty("]]");
-  expect({ a: 1 }).not.toHaveProperty("[]]");
-  expect({ a: 1 }).not.toHaveProperty("[[]");
-  expect({ a: 1 }).not.toHaveProperty(".]");
-  expect({ a: 1 }).not.toHaveProperty(".[");
-  expect({ "": 1 }).toHaveProperty("[.", 1);
-  expect({ a: 1 }).not.toHaveProperty("[.");
-  expect({ a: 1 }).not.toHaveProperty("].");
-  expect({ a: 1 }).not.toHaveProperty("].[");
-  expect({ a: 1 }).not.toHaveProperty("].]");
-  expect({ a: 1 }).not.toHaveProperty("[.]");
-  expect({ a: 1 }).not.toHaveProperty("[.[");
-
-  expect([1]).toHaveProperty("[0]", 1);
-  expect([1]).toHaveProperty("[0][", 1);
-  expect([1]).toHaveProperty("[0]]", 1);
-  expect([1]).toHaveProperty("[0][[", 1);
-  expect([1]).toHaveProperty("[][[[0]", 1);
-  expect([1]).toHaveProperty("[][[[]][[][][.0", 1);
-  expect([1]).toHaveProperty("[][[[]][[][][.[][[][[[][][0", 1);
-  expect([1]).not.toHaveProperty("......1.............", 1);
-  expect([1]).not.toHaveProperty("......0.............", 1);
-  expect([1]).not.toHaveProperty(".0", 1);
-  expect([1]).not.toHaveProperty("0.", 1);
-  expect([{ "": 1 }]).toHaveProperty("0.", 1);
-  expect({ "": { "": 1 } }).toHaveProperty(".", 1);
-  expect({ "": { "": { "": 1 } } }).toHaveProperty("..", 1);
-  expect({ "": { "": { "": 1 } } }).not.toHaveProperty(".", 1);
-  expect({ "": { "": { "": 1 } } }).not.toHaveProperty("...", 1);
-  expect({ "": { "": { "": 1 } } }).not.toHaveProperty("....", 1);
-  expect([1]).toHaveProperty("0.[[[][][]][[[][[]]]]", 1);
-  expect([1]).not.toHaveProperty("[0].", 1);
-  expect([1]).toHaveProperty("0", 1);
-  expect([1]).toHaveProperty("[].0", 1);
-  expect([1]).toHaveProperty("[.0", 1);
-  expect([1]).toHaveProperty("].0", 1);
-  expect([1]).toHaveProperty("0[]][[[]", 1);
-  expect([1]).toHaveProperty("[[]][[[][][0", 1);
-  expect([1]).toHaveProperty("0", 1);
-  expect([1]).toHaveProperty("0.[", 1);
-  expect([1]).not.toHaveProperty("0........[", 1);
-  expect([1]).not.toHaveProperty("0..[", 1);
-  expect([1]).not.toHaveProperty(".0", 1);
-  expect([1]).toHaveProperty("[].0", 1);
-  expect([1]).not.toHaveProperty("[]..0", 1);
-  expect([1]).toHaveProperty("[.][.[[.]]]]].[.[].].]]]]].].].0", 1);
-  expect([1]).not.toHaveProperty("[.][.[[.]]]]].[.[].].]]0]]].].].", 1);
-  expect([1]).toHaveProperty("[.][.[[.]]]]].[.[].].]]0]]].].]", 1);
-  expect([1]).not.toHaveProperty("[.][.[[..]]]]].[.[].].]]0]]].].]", 1);
-  expect([1]).toHaveProperty("[.][.[[.]]]]].[.[].].0.]]]]].].]", 1);
-  expect([1]).not.toHaveProperty("[.][.[[.]]]]].[.[].].0.]]] ]].].]", 1);
-  expect([1]).not.toHaveProperty("0      ", 1);
-  expect([1]).not.toHaveProperty("   0      ", 1);
-  expect([1]).not.toHaveProperty("   0[]      ", 1);
-  expect([1]).not.toHaveProperty("   0]      ", 1);
-  expect([1]).not.toHaveProperty(" .[0]", 1);
-
-  expect({ "": 1 }).not.toHaveProperty(".", 1);
-  expect({ "": 1 }).not.toHaveProperty("]", 1);
-  expect({ "": 1 }).not.toHaveProperty("[", 1);
-  expect({ "": 1 }).toHaveProperty("", 1);
-
-  expect({ "": 1 }).not.toHaveProperty("..", 1);
-  expect({ "": { "": 1 } }).not.toHaveProperty("..", 1);
-  expect([{ "": 1 }]).toHaveProperty("0.", 1);
-  expect([{ "": 1 }]).not.toHaveProperty(".0.", 1);
-  expect({ "": [1] }).toHaveProperty(".0", 1);
-  expect({ "": [1] }).not.toHaveProperty("..0", 1);
-  expect([{ "": 1 }]).not.toHaveProperty("0..", 1);
-  expect([{ "": { "": 1 } }]).toHaveProperty("0..", 1);
-
-  expect([1]).not.toHaveProperty("[0].", 1);
-  expect([1]).not.toHaveProperty("[0][0]", 1);
-  expect({ a: [1] }).toHaveProperty("a[[[[[[[[[0]]]", 1);
-  expect({ "[[[": 0 }).not.toHaveProperty("[[[", 0);
-});
-
-test("toHaveProperty() - with string or array", () => {
-  const a = new Array(["a", "b", "c"]);
-  expect(a).toHaveProperty("0.1", "b");
-  const b = new Array("a", "b", "c");
-  expect({ a: { b: { c: 1 } } }).toHaveProperty(b);
-  const c = {
-    a: { b: 1 },
-    "a.b": 2,
-  };
-  const d = new Array("a.b");
-  expect(c).toHaveProperty(d, 2);
-  const houseForSale = {
-    bath: true,
-    bedrooms: 4,
-    kitchen: {
-      amenities: ["oven", "stove", "washer"],
-      area: 20,
-      wallColor: "white",
-      "nice.oven": true,
-    },
-    livingroom: {
-      amenities: [
-        {
-          couch: [
-            ["large", { dimensions: [20, 20] }],
-            ["small", { dimensions: [10, 10] }],
-          ],
-        },
-      ],
-    },
-    sunroom: "yes",
-    "ceiling.height": 20,
-    "entrance.window": 3,
-    entrance: { window: 5 },
-  };
-  expect(houseForSale).toHaveProperty("entrance.window", 5);
-  expect(houseForSale).toHaveProperty(["entrance", "window"], 5);
-  expect(houseForSale).toHaveProperty(["entrance.window"], 3);
-  expect(houseForSale).toHaveProperty("bath");
-  expect(houseForSale).not.toHaveProperty("jacuzzi");
-  // expect(houseForSale).toHaveProperty("jacuzzi");
-  // expect(houseForSale).not.toHaveProperty("bath");
-  expect(houseForSale).toHaveProperty("bath", true);
-  expect(houseForSale).not.toHaveProperty("bath", false);
-  // expect(houseForSale).toHaveProperty("bath", false);
-  // expect(houseForSale).not.toHaveProperty("bath", true);
-  expect(houseForSale).toHaveProperty("bedrooms", 4);
-  expect(houseForSale).toHaveProperty(["sunroom"], "yes");
-  expect(houseForSale).toHaveProperty("kitchen.area", 20);
-  expect(houseForSale).toHaveProperty("kitchen.amenities", ["oven", "stove", "washer"]);
-  expect(houseForSale).not.toHaveProperty(["kitchen", "area"], 21);
-  expect(houseForSale).toHaveProperty(["kitchen", "area"], 20);
-  expect(houseForSale).not.toHaveProperty(["kitchen", "area"], 29);
-  expect(houseForSale).toHaveProperty(["kitchen", "amenities"], ["oven", "stove", "washer"]);
-  expect(houseForSale).toHaveProperty("kitchen.amenities[2]", "washer");
-  expect(houseForSale).toHaveProperty(["kitchen", "amenities", 1], "stove");
-  expect(houseForSale).toHaveProperty(["kitchen", "amenities", 0], "oven");
-  expect(houseForSale).toHaveProperty("livingroom.amenities[0].couch[0][1].dimensions[0]", 20);
-  expect(houseForSale).toHaveProperty(["kitchen", "nice.oven"]);
-  expect(houseForSale).not.toHaveProperty(["kitchen", "open"]);
-  expect(houseForSale).toHaveProperty(["ceiling.height"], 20);
-  expect({ a: { b: 1 } }).toHaveProperty("a.b");
-  expect({ a: [2, 3, 4] }).toHaveProperty("a.0");
-  expect({ a: [2, 3, 4] }).toHaveProperty("a.1");
-  expect({ a: [2, 3, 4] }).toHaveProperty("a.2");
-  expect({ a: [2, 3, 4] }).toHaveProperty("a[1]");
-  expect([2, 3, 4]).toHaveProperty("1");
-  expect([2, 3, 4]).toHaveProperty("[1]");
-  expect([2, [6, 9], 4]).toHaveProperty("1.1");
-  expect([2, [6, 9], 4]).toHaveProperty("1[1]");
-  expect([2, [6, 9], 4]).toHaveProperty("[1].1");
-  expect([2, [6, 9], 4]).toHaveProperty("[1][1]");
-  expect([2, [6, 9], 4]).toHaveProperty([0], 2);
-  expect({ a: { b: 1 } }).toHaveProperty("a.b");
-  expect({ a: [1, 2, [3, { b: 1 }]] }).toHaveProperty("a.2.1.b");
-  expect({ a: [1, 2, [3, { b: 1 }]] }).toHaveProperty("a");
-  expect({ a: [1, 2, [3, { b: 1 }]] }).toHaveProperty("a[2][1].b");
-  expect({ a: [1, 2, [3, { b: 1 }]] }).toHaveProperty("a[2][1]");
-  expect({ a: [1, 2, [3, { b: 1 }]] }).not.toHaveProperty("a[2][1].c");
-  expect("test").toHaveProperty("length");
-  expect({}).toHaveProperty("constructor");
-  expect({}).toHaveProperty("constructor.name");
-  expect({}).toHaveProperty("constructor.name", "Object");
-  expect(new Date()).toHaveProperty("getTime");
-});
-
-test("toHaveProperty() - all", () => {
-  expect({ a: 1 }).toHaveProperty("a");
-  expect({ a: 1 }).toHaveProperty("a", 1);
-  expect({ a: 1 }).not.toHaveProperty("b");
-  expect({ a: 1 }).not.toHaveProperty("a", 2);
-
-  // test with object with property "a" with all types of values (including undefined)
-  expect({ a: undefined }).toHaveProperty("a");
-  expect({ a: null }).toHaveProperty("a");
-  expect({ a: 0 }).toHaveProperty("a");
-  expect({ a: false }).toHaveProperty("a");
-  expect({ a: "" }).toHaveProperty("a");
-  expect({ a: {} }).toHaveProperty("a");
-  expect({ a: [] }).toHaveProperty("a");
-  expect({ a: () => {} }).toHaveProperty("a");
-
-  // test with object with property "a" with all types of values (including undefined)
-  expect({ a: undefined }).toHaveProperty("a", undefined);
-  expect({ a: null }).toHaveProperty("a", null);
-  expect({ a: 0 }).toHaveProperty("a", 0);
-  expect({ a: false }).toHaveProperty("a", false);
-  expect({ a: "" }).toHaveProperty("a", "");
-  expect({ a: {} }).toHaveProperty("a", {});
-  expect({ a: [] }).toHaveProperty("a", []);
-  expect({ a: () => {} }).not.toHaveProperty("a", () => {});
-
-  // test with object with property "a" with all types of values (including undefined)
-
-  expect({ a: undefined }).not.toHaveProperty("a", null);
-  expect({ a: null }).not.toHaveProperty("a", undefined);
-  expect({ a: 0 }).not.toHaveProperty("a", null);
-  expect({ a: false }).not.toHaveProperty("a", null);
-  expect({ a: "" }).not.toHaveProperty("a", null);
-  expect({ a: {} }).not.toHaveProperty("a", null);
-  expect({ a: [] }).not.toHaveProperty("a", null);
-  expect({ a: () => {} }).not.toHaveProperty("a", null);
-
-  expect({ a: undefined }).not.toHaveProperty("a", 0);
-  expect({ a: null }).not.toHaveProperty("a", 0);
-  expect({ a: 0 }).not.toHaveProperty("a", 1);
-  expect({ a: false }).not.toHaveProperty("a", 0);
-  expect({ a: "" }).not.toHaveProperty("a", 0);
-  expect({ a: {} }).not.toHaveProperty("a", 0);
-  expect({ a: [] }).not.toHaveProperty("a", 0);
-  expect({ a: () => {} }).not.toHaveProperty("a", 0);
-
-  expect({ a: undefined }).not.toHaveProperty("a", false);
-  expect({ a: null }).not.toHaveProperty("a", false);
-  expect({ a: 0 }).not.toHaveProperty("a", false);
-  expect({ a: false }).not.toHaveProperty("a", true);
-  expect({ a: "" }).not.toHaveProperty("a", false);
-  expect({ a: {} }).not.toHaveProperty("a", false);
-  expect({ a: [] }).not.toHaveProperty("a", false);
-  expect({ a: () => {} }).not.toHaveProperty("a", false);
-
-  expect({ a: undefined }).not.toHaveProperty("a", "");
-  expect({ a: null }).not.toHaveProperty("a", "");
-  expect({ a: 0 }).not.toHaveProperty("a", "");
-  expect({ a: false }).not.toHaveProperty("a", "");
-  expect({ a: "" }).not.toHaveProperty("a", "a");
-  expect({ a: {} }).not.toHaveProperty("a", "");
-  expect({ a: [] }).not.toHaveProperty("a", "");
-  expect({ a: () => {} }).not.toHaveProperty("a", "");
-
-  expect({ a: undefined }).not.toHaveProperty("a", {});
-  expect({ a: null }).not.toHaveProperty("a", {});
-  expect({ a: 0 }).not.toHaveProperty("a", {});
-  expect({ a: false }).not.toHaveProperty("a", {});
-  expect({ a: "" }).not.toHaveProperty("a", {});
-  expect({ a: {} }).not.toHaveProperty("a", { a: 1 });
-  expect({ a: [] }).not.toHaveProperty("a", {});
-  expect({ a: () => {} }).not.toHaveProperty("a", {});
-
-  // test object with property "a" with value set, map, string
-  expect({ a: new Set([1, 2, 3]) }).toHaveProperty("a", new Set([3, 2, 1]));
-  expect({ a: new Map([{ a: 1 }, { b: 2 }, { c: 3 }]) }).toHaveProperty("a", new Map([{ c: 3 }, { b: 2 }, { a: 1 }]));
-  expect({ a: new String("a") }).toHaveProperty("a", new String("a"));
-  expect({ a: new String("a") }).not.toHaveProperty("a", "a");
-  expect({ a: new String("a") }).not.toHaveProperty("a", "b");
-  expect({ a: new String("a") }).not.toHaveProperty("a", new String("b"));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Number(1));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Boolean(true));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Boolean(false));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Object());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Function());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Date());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new RegExp());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Error());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Promise(() => {}));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new WeakSet());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new WeakMap());
-  expect({ a: new String("a") }).not.toHaveProperty("a", Symbol("a"));
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Int8Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Uint8Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Uint8ClampedArray());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Int16Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Uint16Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Int32Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Uint32Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Float32Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new Float64Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new BigInt64Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new BigUint64Array());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new ArrayBuffer());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new SharedArrayBuffer());
-  expect({ a: new String("a") }).not.toHaveProperty("a", new DataView(new ArrayBuffer(1)));
-
-  // test property equality with sets, maps, objects, arrays, and String
-  expect({ a: new Set([1, 2, 3]) }).toHaveProperty("a", new Set([1, 2, 3]));
-  expect({ a: new Map([{ a: 1 }, { b: 2 }, { c: 3 }]) }).toHaveProperty("a", new Map([{ a: 1 }, { b: 2 }, { c: 3 }]));
-  expect({ a: { a: 1, b: 2, c: 3 } }).toHaveProperty("a", { a: 1, b: 2, c: 3 });
-  expect({ a: [1, 2, 3] }).toHaveProperty("a", [1, 2, 3]);
-  expect({ a: "a" }).toHaveProperty("a", "a");
-  expect({ a: new String("a") }).toHaveProperty("a", new String("a"));
-  expect({ a: new String("a") }).not.toHaveProperty("a", "a");
-});
-
-test("toHaveProperty() - null or undefined", () => {
-  expect(null).not.toHaveProperty("length");
-  expect(undefined).not.toHaveProperty("length");
-});
-
-test("toBe()", () => {
-  const a = 1;
-  const b = 1;
-  expect(a).toBe(a);
-  expect(a).toBe(b);
-  expect(a).toBe(1);
-  expect(1).toBe(a);
-  expect(b).toBe(a);
-
-  const c = { a: 1 };
-  const d = { a: 1 };
-  expect(c).toBe(c);
-  expect(c).not.toBe(d);
-  expect(c).not.toBe({ a: 1 });
-  expect({ a: 1 }).not.toBe(c);
-  expect(d).not.toBe(c);
-
-  expect(1).toBe(1);
-  // expect(1).not.toBe(1);
-
-  expect(1).not.toBe(2);
-  expect(1).not.toBe("1");
-  expect("hello test").toBe("hello test");
-  expect("hello test").not.toBe("hello test2");
-});
-
-test("toHaveLength()", () => {
-  expect({ length: Number.MAX_SAFE_INTEGER }).toHaveLength(Number.MAX_SAFE_INTEGER);
-  expect("123").toHaveLength(3);
-  expect([1, 2, 3]).toHaveLength(3);
-  expect([1, 2, 3]).not.toHaveLength(2);
-  expect("123").not.toHaveLength(2);
-  expect({ length: 3 }).toHaveLength(3);
-  expect({ length: 3 }).not.toHaveLength(2);
-  expect({ length: 3 }).not.toHaveLength(Number.MAX_SAFE_INTEGER);
-  expect({ length: Number.MAX_SAFE_INTEGER }).not.toHaveLength(Number.MAX_SAFE_INTEGER - 1);
-  expect({ length: 3.3 }).not.toHaveLength(3);
-  expect("123").not.toHaveLength(-0);
-});
-
-test("toHaveLength() extended", () => {
-  // Headers
-  expect(new Headers()).toHaveLength(0);
-  expect(new Headers({ a: "1" })).toHaveLength(1);
-
-  // FormData
-  const form = new FormData();
-  expect(form).toHaveLength(0);
-  form.append("a", "1");
-  expect(form).toHaveLength(1);
-
-  // URLSearchParams
-  expect(new URLSearchParams()).toHaveLength(0);
-  expect(new URLSearchParams("a=1")).toHaveLength(1);
-  expect(new URLSearchParams([["a", "1"]])).toHaveLength(1);
-
-  // files
-  const thisFile = Bun.file(import.meta.path);
-  const thisFileSize = thisFile.size;
-
-  expect(thisFile).toHaveLength(thisFileSize);
-  expect(thisFile).toHaveLength(Bun.file(import.meta.path).size);
-
-  // empty file should have length 0
-  writeFileSync("/tmp/empty.txt", "");
-  expect(Bun.file("/tmp/empty.txt")).toHaveLength(0);
-
-  // if a file doesn't exist, it should throw (not return 0 size)
-  expect(() => expect(Bun.file("/does-not-exist/file.txt")).toHaveLength(0)).toThrow();
-
-  // Blob
-  expect(new Blob([1, 2, 3])).toHaveLength(3);
-  expect(new Blob()).toHaveLength(0);
-
-  // Set
-  expect(new Set()).toHaveLength(0);
-  expect(new Set([1, 2, 3])).toHaveLength(3);
-
-  // Map
-  expect(new Map()).toHaveLength(0);
-  expect(new Map([["a", 1]])).toHaveLength(1);
-
-  // WeakMap
-  expect(new WeakMap([[globalThis, 1]])).toHaveLength(1);
-});
-
-test("toContain()", () => {
-  const s1 = new String("123");
-  expect(s1).not.toContain("12");
-  const s2 = "123";
-  expect(s2).toContain("12");
-
-  expect("test").toContain("es");
-  expect("test").toContain("est");
-  // expect("test").not.toContain("test");
-  expect(["test", "es"]).toContain("es");
-  expect("").toContain("");
-  expect([""]).toContain("");
-
-  expect(["lemon", "lime"]).not.toContain("orange");
-  expect("citrus fruits").toContain("fruit");
-
-  const a = new Uint16Array([1, 2, 3]);
-  expect(a).toContain(2);
-  expect(a).not.toContain(4);
-  expect([2, "2335", 5, true, false, null, undefined]).toContain(5);
-  expect([2, "2335", 5, true, false, null, undefined]).toContain("2335");
-  expect([2, "2335", 5, true, false, null, undefined]).toContain(true);
-  expect([2, "2335", 5, true, false, null, undefined]).toContain(false);
-  expect([2, "2335", 5, true, false, null, undefined]).toContain(null);
-  expect([2, "2335", 5, true, false, null, undefined]).toContain(undefined);
-  expect([2, "2335", 5, true, false, null, undefined]).not.toContain(3);
-  expect([2, "2335", 5, true, false, null, undefined]).not.not.not.toContain(3);
-
-  // expect([4, 5, 6]).not.toContain(5);
-
-  expect([]).not.toContain([]);
-});
-
-test("toBeEven()", () => {
-  expect(1).not.toBeEven();
-  expect(2).toBeEven();
-  expect(3).not.toBeEven();
-  expect(3.1).not.toBeEven();
-  expect(2.1).not.toBeEven();
-  expect(4).toBeEven();
-  expect(5).not.toBeEven();
-  expect(6).toBeEven();
-  expect(0).toBeEven();
-  expect(-8).toBeEven();
-  expect(-0).toBeEven();
-  expect(NaN).not.toBeEven();
-  expect([]).not.toBeEven();
-  expect([1, 2]).not.toBeEven();
-  expect({}).not.toBeEven();
-  expect(() => {}).not.toBeEven();
-  expect("").not.toBeEven();
-  expect("string").not.toBeEven();
-  expect(undefined).not.toBeEven();
-  expect(Math.floor(Date.now() / 1000) * 2).toBeEven(); // Slight fuzz by using timestamp times 2
-  expect(Math.floor(Date.now() / 1000) * 4 - 1).not.toBeEven();
-  expect(4.0e1).toBeEven();
-  expect(6.2e1).toBeEven();
-  expect(6.3e1).not.toBeEven();
-  expect(6.33e1).not.toBeEven();
-  expect(3.3e-1).not.toBeEven(); //throw
-  expect(0.3).not.toBeEven(); //throw
-  expect(0.4).not.toBeEven();
-  expect(1).not.toBeEven();
-  expect(0).toBeEven();
-  expect(2.0).toBeEven();
-  expect(NaN).not.toBeEven();
-  expect(2n).toBeEven(); // BigInt at this time not supported in jest-extended
-  expect(3n).not.toBeEven();
-  expect(9007199254740990).toBeEven(); // manual typical max safe -1 // not int?
-  expect(9007199254740990n).toBeEven(); // manual typical max safe -1 as bigint
-  expect(Number.MAX_SAFE_INTEGER - 1).toBeEven(); // not int?
-  expect(Number.MAX_SAFE_INTEGER).not.toBeEven();
-  expect(BigInt(Number.MAX_SAFE_INTEGER) - 1n).toBeEven();
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeEven();
-  expect(BigInt(Number.MAX_VALUE - 1)).toBeEven();
-  expect(Number.MIN_SAFE_INTEGER + 1).toBeEven(); // not int?
-  expect(Number.MIN_SAFE_INTEGER).not.toBeEven();
-  expect(BigInt(Number.MIN_SAFE_INTEGER) + 1n).toBeEven();
-  expect(BigInt(Number.MIN_SAFE_INTEGER)).not.toBeEven();
-  expect(4 / Number.NEGATIVE_INFINITY).toBeEven(); // as in IEEE-754: + / -inf => neg zero
-  expect(5 / Number.NEGATIVE_INFINITY).toBeEven();
-  expect(-7 / Number.NEGATIVE_INFINITY).toBeEven(); // as in IEEE-754: - / -inf => zero
-  expect(-8 / Number.NEGATIVE_INFINITY).toBeEven();
-  expect(new WebAssembly.Global({ value: "i32", mutable: false }, 4).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "i32", mutable: false }, 3).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "i32", mutable: true }, 2).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "i32", mutable: true }, 1).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "i64", mutable: true }, -9223372036854775808n).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "i64", mutable: false }, -9223372036854775808n).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "i64", mutable: true }, 9223372036854775807n).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "i64", mutable: false }, 9223372036854775807n).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 42.0).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 42.0).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 42.0).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 42.0).value).toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 43.0).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 43.0).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 43.0).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 43.0).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 4.3).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 4.3).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 4.3).value).not.toBeEven();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 4.3).value).not.toBeEven();
-  // did not seem to support SIMD v128 type yet (which is not in W3C specs for JS but is a valid global type)
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:false}, -170141183460469231731687303715884105728n).value).toBeEven();
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:true}, -170141183460469231731687303715884105728n).value).toBeEven();
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:true}, 170141183460469231731687303715884105727n).value).not.toBeEven();
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:false}, 170141183460469231731687303715884105727n).value).not.toBeEven();
-  // FUTURE: with uintv128: expect(new WebAssembly.Global({value:'v128', mutable:false}, 340282366920938463463374607431768211456n).value).toThrow();
-});
-
-test("toBeTruthy()", () => {
-  expect("test").toBeTruthy();
-  expect(true).toBeTruthy();
-  expect(1).toBeTruthy();
-  expect({}).toBeTruthy();
-  expect([]).toBeTruthy();
-  expect(() => {}).toBeTruthy();
-  // expect(() => {}).not.toBeTruthy();
-  expect(0.5).toBeTruthy();
-  expect(new Map()).toBeTruthy();
-
-  expect("").not.toBeTruthy();
-  expect(0).not.toBeTruthy();
-  expect(-0).not.toBeTruthy();
-  expect(NaN).not.toBeTruthy();
-  expect(0n).not.toBeTruthy();
-  expect(0.0e1).not.toBeTruthy();
-  expect(false).not.toBeTruthy();
-  expect(null).not.toBeTruthy();
-  expect(undefined).not.toBeTruthy();
-});
-
-test("toBeUndefined()", () => {
-  expect(undefined).toBeUndefined();
-  // expect(undefined).not.toBeUndefined();
-
-  expect(null).not.toBeUndefined();
-  expect(null).not.not.not.toBeUndefined();
-  expect(0).not.toBeUndefined();
-  expect("hello defined").not.toBeUndefined();
-});
-
-test("toBeNaN()", () => {
-  expect(NaN).toBeNaN();
-  // expect(NaN).not.toBeNaN();
-
-  expect(0).not.toBeNaN();
-  expect("hello not NaN").not.toBeNaN();
-});
-
-test("toBeNull()", () => {
-  expect(null).toBeNull();
-  // expect(null).not.toBeNull();
-
-  expect(undefined).not.toBeNull();
-  expect(0).not.toBeNull();
-  expect("hello not null").not.toBeNull();
-});
-
-test("toBeDefined()", () => {
-  expect(0).toBeDefined();
-  expect("hello defined").toBeDefined();
-  expect(null).toBeDefined();
-  // expect(null).not.toBeDefined();
-
-  expect(undefined).not.toBeDefined();
-});
-
-test("toBeFalsy()", () => {
-  expect("").toBeFalsy();
-  expect(0).toBeFalsy();
-  expect(-0).toBeFalsy();
-  expect(NaN).toBeFalsy();
-  expect(0n).toBeFalsy();
-  expect(false).toBeFalsy();
-  expect(null).toBeFalsy();
-  expect(undefined).toBeFalsy();
-  // expect(undefined).not.toBeFalsy();
-
-  expect("hello not falsy").not.toBeFalsy();
-  expect("hello not falsy").not.not.not.toBeFalsy();
-  expect(1).not.toBeFalsy();
-  expect(true).not.toBeFalsy();
-  expect({}).not.toBeFalsy();
-  expect([]).not.toBeFalsy();
-  expect(() => {}).not.toBeFalsy();
-});
-
-test("toBeGreaterThan()", () => {
-  expect(3n).toBeGreaterThan(2);
-  expect(Number.MAX_VALUE).not.toBeGreaterThan(Number.MAX_VALUE);
-  expect(1).not.toBeGreaterThan(BigInt(Number.MAX_VALUE));
-  expect(1).not.toBeGreaterThan(Number.MAX_SAFE_INTEGER);
-  expect(1).not.toBeGreaterThan(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(Number.MAX_SAFE_INTEGER).not.toBeGreaterThan(Number.MAX_SAFE_INTEGER);
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeGreaterThan(BigInt(Number.MAX_SAFE_INTEGER));
-
-  expect(Infinity).toBeGreaterThan(-Infinity);
-  expect(-Infinity).not.toBeGreaterThan(Infinity);
-
-  expect(NaN).not.toBeGreaterThan(NaN);
-  expect(NaN).not.toBeGreaterThan(-Infinity);
-
-  expect(10).toBeGreaterThan(9);
-  expect(10).not.toBeGreaterThan(10);
-  expect(10).not.toBeGreaterThan(11);
-  expect(10).not.toBeGreaterThan(Infinity);
-  expect(10).toBeGreaterThan(-Infinity);
-  expect(10).not.toBeGreaterThan(NaN);
-  expect(10).toBeGreaterThan(0);
-  expect(10).toBeGreaterThan(-0);
-  expect(10).toBeGreaterThan(0.1);
-  expect(10).toBeGreaterThan(-0.1);
-  expect(10).toBeGreaterThan(0.9);
-  expect(10).toBeGreaterThan(-0.9);
-  expect(10).toBeGreaterThan(1);
-  expect(10).toBeGreaterThan(-1);
-  // switch the order
-  expect(9).not.toBeGreaterThan(10);
-  expect(10).not.toBeGreaterThan(10);
-  expect(11).toBeGreaterThan(10);
-  expect(Infinity).toBeGreaterThan(10);
-  expect(-Infinity).not.toBeGreaterThan(10);
-  expect(NaN).not.toBeGreaterThan(10);
-  expect(0).not.toBeGreaterThan(10);
-  expect(-0).not.toBeGreaterThan(10);
-  expect(0.1).not.toBeGreaterThan(10);
-  expect(-0.1).not.toBeGreaterThan(10);
-  expect(0.9).not.toBeGreaterThan(10);
-  expect(-0.9).not.toBeGreaterThan(10);
-  expect(1).not.toBeGreaterThan(10);
-  expect(-1).not.toBeGreaterThan(10);
-
-  // same tests but use bigints
-  expect(10n).toBeGreaterThan(9n);
-  expect(10n).not.toBeGreaterThan(10n);
-  expect(10n).not.toBeGreaterThan(11n);
-  expect(10n).not.toBeGreaterThan(Infinity);
-  expect(10n).toBeGreaterThan(-Infinity);
-  expect(10n).not.toBeGreaterThan(NaN);
-  expect(10n).toBeGreaterThan(0n);
-  expect(10n).toBeGreaterThan(-0n);
-  expect(10n).toBeGreaterThan(1n);
-  expect(10n).toBeGreaterThan(-1n);
-  // switch the order
-  expect(9n).not.toBeGreaterThan(10n);
-  expect(10n).not.toBeGreaterThan(10n);
-  expect(11n).toBeGreaterThan(10n);
-  expect(Infinity).toBeGreaterThan(10n);
-  expect(-Infinity).not.toBeGreaterThan(10n);
-  expect(NaN).not.toBeGreaterThan(10n);
-  expect(0n).not.toBeGreaterThan(10n);
-  expect(-0n).not.toBeGreaterThan(10n);
-  expect(1n).not.toBeGreaterThan(10n);
-  expect(-1n).not.toBeGreaterThan(10n);
-
-  // use bigints and numbers
-  expect(10n).toBeGreaterThan(9);
-  expect(10n).not.toBeGreaterThan(10);
-  expect(10n).not.toBeGreaterThan(11);
-  expect(10n).not.toBeGreaterThan(Infinity);
-  expect(10n).toBeGreaterThan(-Infinity);
-  expect(10n).not.toBeGreaterThan(NaN);
-  expect(10n).toBeGreaterThan(0);
-  expect(10n).toBeGreaterThan(-0);
-  expect(10n).toBeGreaterThan(0.1);
-  expect(10n).toBeGreaterThan(-0.1);
-  expect(10n).toBeGreaterThan(0.9);
-  expect(10n).toBeGreaterThan(-0.9);
-  expect(10n).toBeGreaterThan(1);
-  expect(10n).toBeGreaterThan(-1);
-  // switch the order
-  expect(9n).not.toBeGreaterThan(10);
-  expect(10n).not.toBeGreaterThan(10);
-  expect(11n).toBeGreaterThan(10);
-  expect(Infinity).toBeGreaterThan(10n);
-  expect(-Infinity).not.toBeGreaterThan(10n);
-  expect(NaN).not.toBeGreaterThan(10n);
-  expect(0n).not.toBeGreaterThan(10);
-  expect(-0n).not.toBeGreaterThan(10);
-  expect(1n).not.toBeGreaterThan(10);
-  expect(-1n).not.toBeGreaterThan(10);
-
-  expect(1n).not.toBeGreaterThan(1);
-  expect(1n).not.toBeGreaterThan(Number.MAX_SAFE_INTEGER);
-  expect(1n).not.toBeGreaterThan(Number.MAX_VALUE);
-  expect(1).not.toBeGreaterThan(1n);
-  expect(Number.MAX_SAFE_INTEGER).toBeGreaterThan(1n);
-  expect(Number.MAX_VALUE).toBeGreaterThan(1n);
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).toBeGreaterThan(1n);
-  expect(BigInt(Number.MAX_VALUE)).toBeGreaterThan(1n);
-  expect(1n).not.toBeGreaterThan(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(1n).not.toBeGreaterThan(BigInt(Number.MAX_VALUE));
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).toBeGreaterThan(1);
-  expect(BigInt(Number.MAX_VALUE)).toBeGreaterThan(1);
-  expect(1).not.toBeGreaterThan(BigInt(Number.MAX_SAFE_INTEGER));
-});
-
-test("toBeGreaterThanOrEqual()", () => {
-  expect(Number.MAX_VALUE).toBeGreaterThanOrEqual(Number.MAX_VALUE);
-  expect(1).not.toBeGreaterThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(1).not.toBeGreaterThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(1).not.toBeGreaterThanOrEqual(BigInt(Number.MAX_VALUE));
-  expect(Number.MAX_SAFE_INTEGER).toBeGreaterThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).toBeGreaterThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-
-  expect(Infinity).toBeGreaterThanOrEqual(-Infinity);
-  expect(-Infinity).not.toBeGreaterThanOrEqual(Infinity);
-
-  expect(NaN).not.toBeGreaterThanOrEqual(NaN);
-  expect(NaN).not.toBeGreaterThanOrEqual(-Infinity);
-
-  expect(10).toBeGreaterThanOrEqual(9);
-  expect(10).toBeGreaterThanOrEqual(10);
-  expect(10).not.toBeGreaterThanOrEqual(11);
-  expect(10).not.toBeGreaterThanOrEqual(Infinity);
-  expect(10).toBeGreaterThanOrEqual(-Infinity);
-  expect(10).not.toBeGreaterThanOrEqual(NaN);
-  expect(10).toBeGreaterThanOrEqual(0);
-  expect(10).toBeGreaterThanOrEqual(-0);
-  expect(10).toBeGreaterThanOrEqual(0.1);
-  expect(10).toBeGreaterThanOrEqual(-0.1);
-  expect(10).toBeGreaterThanOrEqual(0.9);
-  expect(10).toBeGreaterThanOrEqual(-0.9);
-  expect(10).toBeGreaterThanOrEqual(1);
-  expect(10).toBeGreaterThanOrEqual(-1);
-  // switch the order
-  expect(9).not.toBeGreaterThanOrEqual(10);
-  expect(10).toBeGreaterThanOrEqual(10);
-  expect(11).toBeGreaterThanOrEqual(10);
-  expect(Infinity).toBeGreaterThanOrEqual(10);
-  expect(-Infinity).not.toBeGreaterThanOrEqual(10);
-  expect(NaN).not.toBeGreaterThanOrEqual(10);
-  expect(0).not.toBeGreaterThanOrEqual(10);
-  expect(-0).not.toBeGreaterThanOrEqual(10);
-  expect(0.1).not.toBeGreaterThanOrEqual(10);
-  expect(-0.1).not.toBeGreaterThanOrEqual(10);
-  expect(0.9).not.toBeGreaterThanOrEqual(10);
-  expect(-0.9).not.toBeGreaterThanOrEqual(10);
-  expect(1).not.toBeGreaterThanOrEqual(10);
-  expect(-1).not.toBeGreaterThanOrEqual(10);
-
-  // same tests but use bigints
-  expect(10n).toBeGreaterThanOrEqual(9n);
-  expect(10n).toBeGreaterThanOrEqual(10n);
-  expect(10n).not.toBeGreaterThanOrEqual(11n);
-  expect(10n).not.toBeGreaterThanOrEqual(Infinity);
-  expect(10n).toBeGreaterThanOrEqual(-Infinity);
-  expect(10n).not.toBeGreaterThanOrEqual(NaN);
-  expect(10n).toBeGreaterThanOrEqual(0n);
-  expect(10n).toBeGreaterThanOrEqual(-0n);
-  expect(10n).toBeGreaterThanOrEqual(1n);
-  expect(10n).toBeGreaterThanOrEqual(-1n);
-  // switch the order
-  expect(9n).not.toBeGreaterThanOrEqual(10n);
-  expect(10n).toBeGreaterThanOrEqual(10n);
-  expect(11n).toBeGreaterThanOrEqual(10n);
-  expect(Infinity).toBeGreaterThanOrEqual(10n);
-  expect(-Infinity).not.toBeGreaterThanOrEqual(10n);
-  expect(NaN).not.toBeGreaterThanOrEqual(10n);
-  expect(0n).not.toBeGreaterThanOrEqual(10n);
-  expect(-0n).not.toBeGreaterThanOrEqual(10n);
-  expect(1n).not.toBeGreaterThanOrEqual(10n);
-  expect(-1n).not.toBeGreaterThanOrEqual(10n);
-
-  // use bigints and numbers
-  expect(10n).toBeGreaterThanOrEqual(9);
-  expect(10n).toBeGreaterThanOrEqual(10);
-  expect(10n).not.toBeGreaterThanOrEqual(11);
-  expect(10n).not.toBeGreaterThanOrEqual(Infinity);
-  expect(10n).toBeGreaterThanOrEqual(-Infinity);
-  expect(10n).not.toBeGreaterThanOrEqual(NaN);
-  expect(10n).toBeGreaterThanOrEqual(0);
-  expect(10n).toBeGreaterThanOrEqual(-0);
-  expect(10n).toBeGreaterThanOrEqual(0.1);
-  expect(10n).toBeGreaterThanOrEqual(-0.1);
-  expect(10n).toBeGreaterThanOrEqual(0.9);
-  expect(10n).toBeGreaterThanOrEqual(-0.9);
-  expect(10n).toBeGreaterThanOrEqual(1);
-  expect(10n).toBeGreaterThanOrEqual(-1);
-  // switch the order
-  expect(9n).not.toBeGreaterThanOrEqual(10);
-  expect(10n).toBeGreaterThanOrEqual(10);
-  expect(11n).toBeGreaterThanOrEqual(10);
-  expect(Infinity).toBeGreaterThanOrEqual(10n);
-  expect(-Infinity).not.toBeGreaterThanOrEqual(10n);
-  expect(NaN).not.toBeGreaterThanOrEqual(10n);
-  expect(0n).not.toBeGreaterThanOrEqual(10);
-  expect(-0n).not.toBeGreaterThanOrEqual(10);
-  expect(1n).not.toBeGreaterThanOrEqual(10);
-  expect(-1n).not.toBeGreaterThanOrEqual(10);
-
-  expect(1n).toBeGreaterThanOrEqual(1);
-  expect(1n).not.toBeGreaterThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(1n).not.toBeGreaterThanOrEqual(Number.MAX_VALUE);
-  expect(1).toBeGreaterThanOrEqual(1n);
-  expect(Number.MAX_SAFE_INTEGER).toBeGreaterThanOrEqual(1n);
-  expect(Number.MAX_VALUE).toBeGreaterThanOrEqual(1n);
-
-  expect(1).not.toBeGreaterThanOrEqual(BigInt(Number.MAX_VALUE));
-});
-
-test("toBeLessThan()", () => {
-  expect(3n).not.toBeLessThan(2);
-  expect(Number.MAX_VALUE).not.toBeLessThan(Number.MAX_VALUE);
-  expect(1).toBeLessThan(BigInt(Number.MAX_VALUE));
-  expect(1).toBeLessThan(Number.MAX_SAFE_INTEGER);
-  expect(1).toBeLessThan(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(Number.MAX_SAFE_INTEGER).not.toBeLessThan(Number.MAX_SAFE_INTEGER);
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeLessThan(BigInt(Number.MAX_SAFE_INTEGER));
-
-  expect(Number.MAX_VALUE).not.toBeLessThan(BigInt(Number.MAX_VALUE));
-
-  expect(NaN).not.toBeLessThan(NaN);
-  expect(NaN).not.toBeLessThan(-Infinity);
-
-  expect(10).not.toBeLessThan(9);
-  expect(10).not.toBeLessThan(10);
-  expect(10).toBeLessThan(11);
-  expect(10).toBeLessThan(Infinity);
-  expect(10).not.toBeLessThan(-Infinity);
-  expect(10).not.toBeLessThan(NaN);
-  expect(10).not.toBeLessThan(0);
-  expect(10).not.toBeLessThan(-0);
-  expect(10).not.toBeLessThan(0.1);
-  expect(10).not.toBeLessThan(-0.1);
-  expect(10).not.toBeLessThan(0.9);
-  expect(10).not.toBeLessThan(-0.9);
-  expect(10).not.toBeLessThan(1);
-  expect(10).not.toBeLessThan(-1);
-  // switch the order
-  expect(9).toBeLessThan(10);
-  expect(10).not.toBeLessThan(10);
-  expect(11).not.toBeLessThan(10);
-  expect(Infinity).not.toBeLessThan(10);
-  expect(-Infinity).toBeLessThan(10);
-  expect(NaN).not.toBeLessThan(10);
-  expect(0).toBeLessThan(10);
-  expect(-0).toBeLessThan(10);
-  expect(0.1).toBeLessThan(10);
-  expect(-0.1).toBeLessThan(10);
-  expect(0.9).toBeLessThan(10);
-  expect(-0.9).toBeLessThan(10);
-  expect(1).toBeLessThan(10);
-  expect(-1).toBeLessThan(10);
-
-  // same tests but use bigints
-  expect(10n).not.toBeLessThan(9n);
-  expect(10n).not.toBeLessThan(10n);
-  expect(10n).toBeLessThan(11n);
-  expect(10n).toBeLessThan(Infinity);
-  expect(10n).not.toBeLessThan(-Infinity);
-  expect(10n).not.toBeLessThan(NaN);
-  expect(10n).not.toBeLessThan(0n);
-  expect(10n).not.toBeLessThan(-0n);
-  expect(10n).not.toBeLessThan(1n);
-  expect(10n).not.toBeLessThan(-1n);
-  // switch the order
-  expect(9n).toBeLessThan(10n);
-  expect(10n).not.toBeLessThan(10n);
-  expect(11n).not.toBeLessThan(10n);
-  expect(Infinity).not.toBeLessThan(10n);
-  expect(-Infinity).toBeLessThan(10n);
-  expect(NaN).not.toBeLessThan(10n);
-  expect(0n).toBeLessThan(10n);
-  expect(-0n).toBeLessThan(10n);
-  expect(1n).toBeLessThan(10n);
-  expect(-1n).toBeLessThan(10n);
-
-  // use bigints and numbers
-  expect(10n).not.toBeLessThan(9);
-  expect(10n).not.toBeLessThan(10);
-  expect(10n).toBeLessThan(11);
-  expect(10n).toBeLessThan(Infinity);
-  expect(10n).not.toBeLessThan(-Infinity);
-  expect(10n).not.toBeLessThan(NaN);
-  expect(10n).not.toBeLessThan(0);
-  expect(10n).not.toBeLessThan(-0);
-  expect(10n).not.toBeLessThan(0.1);
-  expect(10n).not.toBeLessThan(-0.1);
-  expect(10n).not.toBeLessThan(0.9);
-  expect(10n).not.toBeLessThan(-0.9);
-  expect(10n).not.toBeLessThan(1);
-  expect(10n).not.toBeLessThan(-1);
-  // switch the order
-  expect(9n).toBeLessThan(10);
-  expect(10n).not.toBeLessThan(10);
-  expect(11n).not.toBeLessThan(10);
-  expect(Infinity).not.toBeLessThan(10n);
-  expect(-Infinity).toBeLessThan(10n);
-  expect(NaN).not.toBeLessThan(10n);
-  expect(0n).toBeLessThan(10);
-  expect(-0n).toBeLessThan(10);
-  expect(1n).toBeLessThan(10);
-  expect(-1n).toBeLessThan(10);
-
-  expect(1n).not.toBeLessThan(1);
-  expect(1n).toBeLessThan(Number.MAX_SAFE_INTEGER);
-  expect(1n).toBeLessThan(Number.MAX_VALUE);
-  expect(1).not.toBeLessThan(1n);
-  expect(Number.MAX_SAFE_INTEGER).not.toBeLessThan(1n);
-  expect(Number.MAX_VALUE).not.toBeLessThan(1n);
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeLessThan(1n);
-  expect(BigInt(Number.MAX_VALUE)).not.toBeLessThan(1n);
-  expect(1n).toBeLessThan(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(1n).toBeLessThan(BigInt(Number.MAX_VALUE));
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeLessThan(1);
-  expect(BigInt(Number.MAX_VALUE)).not.toBeLessThan(1);
-  expect(1).toBeLessThan(BigInt(Number.MAX_SAFE_INTEGER));
-});
-
-test("toBeLessThanOrEqual()", () => {
-  expect(3n).not.toBeLessThanOrEqual(2);
-  expect(Number.MAX_VALUE).toBeLessThanOrEqual(Number.MAX_VALUE);
-  expect(1).toBeLessThanOrEqual(BigInt(Number.MAX_VALUE));
-  expect(1).toBeLessThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(1).toBeLessThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(Number.MAX_SAFE_INTEGER).toBeLessThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).toBeLessThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-
-  expect(Number.MAX_VALUE).toBeLessThanOrEqual(BigInt(Number.MAX_VALUE));
-  expect(BigInt(Number.MAX_VALUE)).toBeLessThanOrEqual(Number.MAX_VALUE);
-
-  expect(NaN).not.toBeLessThanOrEqual(NaN);
-  expect(NaN).not.toBeLessThanOrEqual(-Infinity);
-
-  expect(10).not.toBeLessThanOrEqual(9);
-  expect(10).toBeLessThanOrEqual(10);
-  expect(10).toBeLessThanOrEqual(11);
-  expect(10).toBeLessThanOrEqual(Infinity);
-  expect(10).not.toBeLessThanOrEqual(-Infinity);
-  expect(10).not.toBeLessThanOrEqual(NaN);
-  expect(10).not.toBeLessThanOrEqual(0);
-  expect(10).not.toBeLessThanOrEqual(-0);
-  expect(10).not.toBeLessThanOrEqual(0.1);
-  expect(10).not.toBeLessThanOrEqual(-0.1);
-  expect(10).not.toBeLessThanOrEqual(0.9);
-  expect(10).not.toBeLessThanOrEqual(-0.9);
-  expect(10).not.toBeLessThanOrEqual(1);
-  expect(10).not.toBeLessThanOrEqual(-1);
-  // switch the order
-  expect(9).toBeLessThanOrEqual(10);
-  expect(10).toBeLessThanOrEqual(10);
-  expect(11).not.toBeLessThanOrEqual(10);
-  expect(Infinity).not.toBeLessThanOrEqual(10);
-  expect(-Infinity).toBeLessThanOrEqual(10);
-  expect(NaN).not.toBeLessThanOrEqual(10);
-  expect(0).toBeLessThanOrEqual(10);
-  expect(-0).toBeLessThanOrEqual(10);
-  expect(0.1).toBeLessThanOrEqual(10);
-  expect(-0.1).toBeLessThanOrEqual(10);
-  expect(0.9).toBeLessThanOrEqual(10);
-  expect(-0.9).toBeLessThanOrEqual(10);
-  expect(1).toBeLessThanOrEqual(10);
-  expect(-1).toBeLessThanOrEqual(10);
-
-  // same tests but use bigints
-  expect(10n).not.toBeLessThanOrEqual(9n);
-  expect(10n).toBeLessThanOrEqual(10n);
-  expect(10n).toBeLessThanOrEqual(11n);
-  expect(10n).toBeLessThanOrEqual(Infinity);
-  expect(10n).not.toBeLessThanOrEqual(-Infinity);
-  expect(10n).not.toBeLessThanOrEqual(NaN);
-  expect(10n).not.toBeLessThanOrEqual(0n);
-  expect(10n).not.toBeLessThanOrEqual(-0n);
-  expect(10n).not.toBeLessThanOrEqual(1n);
-  expect(10n).not.toBeLessThanOrEqual(-1n);
-  // switch the order
-  expect(9n).toBeLessThanOrEqual(10n);
-  expect(10n).toBeLessThanOrEqual(10n);
-  expect(11n).not.toBeLessThanOrEqual(10n);
-  expect(Infinity).not.toBeLessThanOrEqual(10n);
-  expect(-Infinity).toBeLessThanOrEqual(10n);
-  expect(NaN).not.toBeLessThanOrEqual(10n);
-  expect(0n).toBeLessThanOrEqual(10n);
-  expect(-0n).toBeLessThanOrEqual(10n);
-  expect(1n).toBeLessThanOrEqual(10n);
-  expect(-1n).toBeLessThanOrEqual(10n);
-
-  // use bigints and numbers
-  expect(10n).not.toBeLessThanOrEqual(9);
-  expect(10n).toBeLessThanOrEqual(10);
-  expect(10n).toBeLessThanOrEqual(11);
-  expect(10n).toBeLessThanOrEqual(Infinity);
-  expect(10n).not.toBeLessThanOrEqual(-Infinity);
-  expect(10n).not.toBeLessThanOrEqual(NaN);
-  expect(10n).not.toBeLessThanOrEqual(0);
-  expect(10n).not.toBeLessThanOrEqual(-0);
-  expect(10n).not.toBeLessThanOrEqual(0.1);
-  expect(10n).not.toBeLessThanOrEqual(-0.1);
-  expect(10n).not.toBeLessThanOrEqual(0.9);
-  expect(10n).not.toBeLessThanOrEqual(-0.9);
-  expect(10n).not.toBeLessThanOrEqual(1);
-  expect(10n).not.toBeLessThanOrEqual(-1);
-  // switch the order
-  expect(9n).toBeLessThanOrEqual(10);
-  expect(10n).toBeLessThanOrEqual(10);
-  expect(11n).not.toBeLessThanOrEqual(10);
-  expect(Infinity).not.toBeLessThanOrEqual(10n);
-  expect(-Infinity).toBeLessThanOrEqual(10n);
-  expect(NaN).not.toBeLessThanOrEqual(10n);
-  expect(0n).toBeLessThanOrEqual(10);
-  expect(-0n).toBeLessThanOrEqual(10);
-  expect(1n).toBeLessThanOrEqual(10);
-  expect(-1n).toBeLessThanOrEqual(10);
-
-  expect(1n).toBeLessThanOrEqual(1);
-  expect(1n).toBeLessThanOrEqual(Number.MAX_SAFE_INTEGER);
-  expect(1n).toBeLessThanOrEqual(Number.MAX_VALUE);
-  expect(1).toBeLessThanOrEqual(1n);
-  expect(Number.MAX_SAFE_INTEGER).not.toBeLessThanOrEqual(1n);
-  expect(Number.MAX_VALUE).not.toBeLessThanOrEqual(1n);
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeLessThanOrEqual(1n);
-  expect(BigInt(Number.MAX_VALUE)).not.toBeLessThanOrEqual(1n);
-  expect(1n).toBeLessThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-  expect(1n).toBeLessThanOrEqual(BigInt(Number.MAX_VALUE));
-
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).not.toBeLessThanOrEqual(1);
-  expect(BigInt(Number.MAX_VALUE)).not.toBeLessThanOrEqual(1);
-  expect(1).toBeLessThanOrEqual(BigInt(Number.MAX_SAFE_INTEGER));
-});
-
-test("toBeOdd()", () => {
-  expect(1).toBeOdd();
-  expect(2).not.toBeOdd();
-  expect(3).toBeOdd();
-  expect(3.1).not.toBeOdd();
-  expect(2.1).not.toBeOdd();
-  expect(4).not.toBeOdd();
-  expect(5).toBeOdd();
-  expect(6).not.toBeOdd();
-  expect(0).not.toBeOdd();
-  expect(-8).not.toBeOdd();
-  expect(-0).not.toBeOdd();
-  expect(NaN).not.toBeOdd();
-  expect([]).not.toBeOdd();
-  // SHOULD FAIL: expect([]).toBeOdd();
-  expect([1, 2]).not.toBeOdd();
-  expect({}).not.toBeOdd();
-  expect(() => {}).not.toBeOdd();
-  expect("").not.toBeOdd();
-  expect("string").not.toBeOdd();
-  expect(undefined).not.toBeOdd();
-  expect(Math.floor(Date.now() / 1000) * 2 - 1).toBeOdd(); // Slight fuzz by using timestamp times 2
-  expect(Math.floor(Date.now() / 1000) * 4 - 1).toBeOdd();
-  expect(4.0e1).not.toBeOdd();
-  expect(6.2e1).not.toBeOdd();
-  expect(6.3e1).toBeOdd();
-  expect(6.33e1).not.toBeOdd();
-  expect(3.2e-3).not.toBeOdd();
-  expect(0.3).not.toBeOdd();
-  expect(0.4).not.toBeOdd();
-  expect(1).toBeOdd();
-  expect(0).not.toBeOdd();
-  expect(2.0).not.toBeOdd();
-  expect(NaN).not.toBeOdd();
-  expect(2n).not.toBeOdd(); // BigInt at this time not supported in jest-extended
-  expect(3n).toBeOdd();
-  expect(9007199254740990).not.toBeOdd(); // manual typical max safe -1
-  expect(9007199254740991).toBeOdd();
-  expect(9007199254740990n).not.toBeOdd(); // manual typical max safe -1 as bigint
-  expect(9007199254740991n).toBeOdd();
-  expect(Number.MAX_SAFE_INTEGER - 1).not.toBeOdd();
-  expect(Number.MAX_SAFE_INTEGER).toBeOdd();
-  expect(BigInt(Number.MAX_SAFE_INTEGER) - 1n).not.toBeOdd();
-  expect(BigInt(Number.MAX_SAFE_INTEGER)).toBeOdd();
-  expect(Number.MIN_SAFE_INTEGER + 1).not.toBeOdd();
-  expect(Number.MIN_SAFE_INTEGER).toBeOdd();
-  expect(BigInt(Number.MIN_SAFE_INTEGER) + 1n).not.toBeOdd();
-  expect(BigInt(Number.MIN_SAFE_INTEGER)).toBeOdd();
-  expect(4 / Number.NEGATIVE_INFINITY).not.toBeOdd(); // in IEEE-754: + / -inf => neg zero
-  expect(5 / Number.NEGATIVE_INFINITY).not.toBeOdd();
-  expect(-7 / Number.NEGATIVE_INFINITY).not.toBeOdd(); // in IEEE-754: - / -inf => zero
-  expect(-8 / Number.NEGATIVE_INFINITY).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "i32", mutable: false }, 4).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "i32", mutable: false }, 3).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "i32", mutable: true }, 2).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "i32", mutable: true }, 1).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "i64", mutable: true }, -9223372036854775808n).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "i64", mutable: false }, -9223372036854775808n).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "i64", mutable: true }, 9223372036854775807n).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "i64", mutable: false }, 9223372036854775807n).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 42.0).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 42.0).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 42.0).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 42.0).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 43.0).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 43.0).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 43.0).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 43.0).value).toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: true }, 4.3).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f32", mutable: false }, 4.3).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: true }, 4.3).value).not.toBeOdd();
-  expect(new WebAssembly.Global({ value: "f64", mutable: false }, 4.3).value).not.toBeOdd();
-  // did not seem to support SIMD v128 type yet
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:false}, 42).value).not.toBeOdd();
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:true}, 42).value).not.toBeOdd();
-  // FUTURE: expect(new WebAssembly.Global({value:'v128', mutable:true}, 43).value).toBeOdd();
-});
-
-describe("toMatchObject", () => {
-  test("with Bun.deepMatch", () => {
-    expect(Bun.deepMatch({ a: 1, b: 2 }, { a: 1 })).toBe(false);
-    expect(Bun.deepMatch({ a: 1 }, { a: 1, b: 2 })).toBe(true);
-  });
-  test("with expect matcher", () => {
-    const f = Symbol.for("foo");
-    const b = Symbol.for("bar");
-
-    class Number2 extends Number {
-      constructor(value) {
-        super(value);
-      }
-    }
-    class Number3 extends Number2 {
-      constructor(value) {
-        super(value);
-      }
-    }
-
-    class Boolean2 extends Boolean {
-      constructor(value) {
-        super(value);
-      }
-    }
-    expect({ [f]: 2 }).toMatchObject({ [f]: 2 });
-    expect({ [f]: 2 }).toMatchObject({ [f]: expect.anything() });
-    expect({ [f]: new Date() }).toMatchObject({ [f]: expect.any(Date) });
-    expect({ [f]: new Date() }).not.toMatchObject({ [f]: expect.any(RegExp) });
-    expect({ [f]: 3 }).not.toMatchObject({ [f]: 5 });
-    expect({ [f]: 3 }).not.toMatchObject({ [b]: 3 });
-    expect({}).toMatchObject({});
-    expect([5]).toMatchObject([5]);
-    expect([5]).not.toMatchObject([4]);
-    expect(() => {
-      expect({}).toMatchObject();
-    }).toThrow();
-    expect(() => {
-      expect(true).toMatchObject(true);
-    }).toThrow();
-    expect(() => {
-      expect(true).toMatchObject(true);
-    }).toThrow();
-    expect(() => {
-      expect(1).toMatchObject(1);
-    }).toThrow();
-    expect(() => {
-      expect("a").toMatchObject("a");
-    }).toThrow();
-    expect(() => {
-      expect(null).toMatchObject(null);
-    }).toThrow();
-    expect(() => {
-      expect(undefined).toMatchObject(undefined);
-    }).toThrow();
-    expect(() => {
-      expect(Symbol()).toMatchObject(Symbol());
-    }).toThrow();
-    expect(() => {
-      expect(BigInt(1)).toMatchObject(BigInt(1));
-    }).toThrow();
-    expect([]).toMatchObject([]);
-    expect([1]).toMatchObject([1]);
-    expect([1, 2]).toMatchObject([1, 2]);
-    expect(() => {
-      expect([1]).toMatchObject([1, 2]);
-    }).toThrow();
-    expect(() => {
-      expect([1, 2]).toMatchObject([1]);
-    }).toThrow();
-    expect([]).toMatchObject({});
-    expect([1]).toMatchObject({});
-    expect([1, 2]).toMatchObject({ 0: 1, 1: 2 });
-    expect([1, 2]).not.toMatchObject({ 0: 2 });
-    expect(() => {
-      expect({}).toMatchObject([]);
-    }).toThrow();
-    expect({ a: 1 }).toMatchObject({});
-    expect({ a: 1 }).toMatchObject({ a: expect.anything() });
-    expect({ a: 1, b: 2 }).toMatchObject({ a: 1 });
-    expect({ a: 1, b: 2 }).toMatchObject({ a: 1, b: 2 });
-    expect({ a: 1, b: 2 }).toMatchObject({ b: 2 });
-    expect({ a: 1, b: 2 }).toMatchObject({ b: 2, a: 1 });
-    expect({ a: 1, b: 2 }).toMatchObject({ a: 1, b: 2 });
-    expect({}).not.toMatchObject({ a: 1 });
-    expect({ a: 89 }).not.toMatchObject({ b: 90 });
-    expect({ a: 1, b: 2 }).not.toMatchObject({ a: 1, b: 3 });
-    expect({ a: 1, b: 2 }).not.toMatchObject({ a: 1, b: 2, c: 4 });
-    expect({ a: new Date(), b: "jj" }).not.toMatchObject({ b: expect.any(Number) });
-    expect({ a: "123" }).not.toMatchObject({ a: expect.stringContaining("4") });
-    class DString extends String {
-      constructor(str) {
-        super(str);
-      }
-    }
-    expect({ a: "hello world" }).toMatchObject({ a: expect.stringContaining("wor") });
-    expect({ a: "hello world" }).not.toMatchObject({ a: expect.stringContaining("wol") });
-    expect({ a: "hello String" }).toMatchObject({ a: expect.stringContaining(new String("Str")) });
-    expect({ a: "hello String" }).not.toMatchObject({ a: expect.stringContaining(new String("Strs")) });
-    expect({ a: "hello derived String" }).toMatchObject({ a: expect.stringContaining(new DString("riv")) });
-    expect({ a: "hello derived String" }).not.toMatchObject({ a: expect.stringContaining(new DString("rivd")) });
-    expect({ a: "hello world" }).toMatchObject({ a: expect.stringMatching("wor") });
-    expect({ a: "hello world" }).not.toMatchObject({ a: expect.stringMatching("word") });
-    expect({ a: "hello world" }).toMatchObject({ a: "hello world" });
-    expect({ a: "hello world" }).toMatchObject({ a: expect.stringMatching(/wor/) });
-    expect({ a: "hello world" }).not.toMatchObject({ a: expect.stringMatching(/word/) });
-    expect({ a: expect.stringMatching("wor") }).toMatchObject({ a: "hello world" });
-    expect({ a: expect.stringMatching("word") }).not.toMatchObject({ a: "hello world" });
-    expect({ a: expect.stringMatching(/wor/) }).toMatchObject({ a: "hello world" });
-    expect({ a: expect.stringMatching(/word/) }).not.toMatchObject({ a: "hello world" });
-    expect({ a: expect.stringMatching(/word/) }).toMatchObject({ a: "hello word" });
-    expect({ a: [1, 2, 3] }).toMatchObject({ a: [1, 2, 3] });
-    expect({ a: [1, 2, 3] }).toMatchObject({ a: [1, 2, 3] });
-    expect({ a: [1, 2, 4] }).not.toMatchObject({ a: [1, 2, 3] });
-
-    expect([]).toMatchObject([]);
-    expect([]).toMatchObject({});
-    expect({}).not.toMatchObject([]);
-    expect({ a: 1 }).toMatchObject({});
-    expect({ a: 1 }).toMatchObject({ a: 1 });
-
-    expect({ a: 1 }).toMatchObject({ a: expect.anything() });
-    expect({ a: null }).not.toMatchObject({ a: expect.anything() });
-    expect({ a: undefined }).not.toMatchObject({ a: expect.anything() });
-
-    expect({ a: new Date() }).toMatchObject({ a: expect.any(Date) });
-    expect({ a: new Date() }).not.toMatchObject({ a: expect.any(RegExp) });
-    expect({ a: new RegExp("a", "g") }).toMatchObject({ a: expect.any(RegExp) });
-    expect({ a: /a/g }).toMatchObject({ a: expect.any(RegExp) });
-
-    expect({
-      first: new Boolean2(false),
-      a: {
-        4: [3, 2, 2],
-        j: new Date(),
-        b: {
-          c: {
-            num: 1,
-            d: {
-              e: {
-                bigint: 123n,
-                f: {
-                  g: {
-                    h: {
-                      i: new Number3(2),
-                      bool: true,
-                    },
-                    compare: "compare",
-                  },
-                },
-                ignore1: 234,
-                ignore2: {
-                  ignore3: 23421,
-                  ignore4: {
-                    ignore5: {
-                      ignore6: "hello",
-                      ignore7: "done",
-                    },
-                  },
-                },
-              },
-            },
-            string1: "hello",
-            string2: "hello",
-            string3: "hello",
-          },
-        },
-      },
-    }).toMatchObject({
-      first: expect.any(Boolean2),
-      a: {
-        4: [3, 2, expect.any(Number)],
-
-        j: expect.any(Date),
-        b: {
-          c: {
-            num: expect.any(Number),
-            string1: expect.anything(),
-            string2: expect.stringContaining("ll"),
-            string3: expect.stringMatching(/ll/),
-            d: {
-              e: {
-                bigint: expect.any(BigInt),
-                f: {
-                  g: {
-                    compare: "compare",
-                    h: {
-                      i: expect.any(Number3),
-                      bool: expect.any(Boolean),
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    });
-
-    var a1 = [1];
-    a1[f] = 99;
-    expect(a1).not.toMatchObject([1]);
-    expect([1]).not.toMatchObject(a1);
-    expect({ 1: 1 }).not.toMatchObject(a1);
-    expect(a1).not.toMatchObject({ 1: 1 });
-    expect(a1).toMatchObject(a1);
-  });
-});
-
 try {
   test("test this doesnt crash");
 } catch (e) {}
